cmf:  - Changed int for signed index to ptrdiff_t for bit equivalent with size_t on Win x64 (affects all files changed in this revision).  - Renamed PredictCorrectSimple to HeunIntegrator.  - Added WaterSoluteIntagrator.cpp to complete last revision

+++ cmf/cmf_core_src/atmosphere/meteorology.cpp
@@ -113,10 +113,10 @@ void cmf::atmosphere::MeteoStation::SetSunshineFraction(cmf::math::timeseries su
 {
 	Sunshine=cmf::math::timeseries(sunshine_duration.begin(),sunshine_duration.step(),sunshine_duration.interpolationpower());
 	cmf::math::Time t=sunshine_duration.begin();
-	for (int i = 0; i < sunshine_duration.size() ; i++)
+	for (ptrdiff_t i = 0; i < sunshine_duration.size() ; i++)
 	{
 		double
-			DOY=t.AsDate().DOY()+t.AsDays()-int(t.AsDays()),
+			DOY=t.AsDate().DOY()+t.AsDays()-ptrdiff_t(t.AsDays()),
 			phi=Latitude*PI/180,	 // Latitude [rad]
 			decl=0.409*sin(2*PI/365*DOY-1.39), // Declination [rad]
 			sunset_angle=acos(-tan(phi)*tan(decl)),	// Sunset hour angle [rad]
@@ -153,7 +153,7 @@ void cmf::atmosphere::MeteoStation::use_for_cell( cmf::upslope::Cell& c )
 {
 	MeteoStation::ptr sh_this;
 	const MeteoStationList& msl =  c.get_project().meteo_stations;
-	for (int i = 0; i < msl.size() ; ++i)
+	for (ptrdiff_t i = 0; i < msl.size() ; ++i)
 	{
 		if (msl[i].get() == this) {
 			MeteoStationReference mref(msl[i],c.get_position());
@@ -167,8 +167,8 @@ void cmf::atmosphere::MeteoStation::use_for_cell( cmf::upslope::Cell& c )
 double cmf::atmosphere::MeteoStationList::calculate_Temp_lapse( cmf::math::Time begin, cmf::math::Time step,cmf::math::Time end )
 {
 	using namespace cmf::math;
-	int n = size();
-	int steps=0;
+	ptrdiff_t n = size();
+	ptrdiff_t steps=0;
 	timeseries temp_lapse(begin,step);
 	double avg_lapse=0;
 	double inv_size = 1./(n-1.);
@@ -237,7 +237,7 @@ cmf::atmosphere::IDW_Meteorology::IDW_Meteorology( const cmf::geometry::point& p
 	point p=position;
 	// Create a vector of distances to the stations
 	num_array dist(stations.size());
-	for (int i = 0; i < stations.size() ; ++i)
+	for (ptrdiff_t i = 0; i < stations.size() ; ++i)
 		dist[i] = p.distanceTo(point(stations[i]->x,stations[i]->y))+z_weight*abs(p.z-stations[i]->z);
 	
 	// Calculate the weight of each station using IDW
@@ -245,7 +245,7 @@ cmf::atmosphere::IDW_Meteorology::IDW_Meteorology( const cmf::geometry::point& p
 	double weightsum = weights_.sum();
 	weights_/=weightsum; // normalize the weights
 	// Create a weight map
-	for (int i = 0; i < stations.size() ; ++i)
+	for (ptrdiff_t i = 0; i < stations.size() ; ++i)
 		weights[stations[i]] = weights_[i];
 }
